% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_transform.R
\name{landsat_crop}
\alias{landsat_crop}
\alias{landsat_mask}
\title{Crop Landsat objects to a boundary}
\usage{
landsat_crop(x, boundary)

landsat_mask(x, boundary)
}
\arguments{
\item{x}{A \link{landsat_scene} or landsat_scene_df (from \link{landsat_load_scenes})}

\item{boundary}{Usually an object of type
\link[sp]{SpatialPolygons-class}, but could also be from the
sf package, a \link[sp]{bbox}, or an \link[raster]{extent}.}
}
\value{
A landsat_scene or landsat_scene_df
}
\description{
This function crops a \link{landsat_scene} or landsat_scene_df
(as generated by \link{landsat_load_scenes}) to a spatial object
describing a boundary. This is usually an object of type
\link[sp]{SpatialPolygons-class}, but could also be from the
sf package, a \link[sp]{bbox}, or an \link[raster]{extent}.
}
