% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_files.R
\name{landsat_list_files}
\alias{landsat_list_files}
\title{List Landsat Files}
\usage{
landsat_list_files(path = ".", full.names = TRUE, recursive = FALSE)
}
\arguments{
\item{path}{The directory where files are located}

\item{full.names}{Use full names in relation to the working directory}

\item{recursive}{Look in all subdirectories.}
}
\value{
A character vector of file locations.
}
\description{
List files in \code{path} that are valid landsat GeoTiff filenames, which
are something like "LT05_L1TP_063013_19960711_20170104_01_T1_B2.TIF".
}
\examples{
landsat_list_files()

}
