% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_scene.R
\name{landsat_load_scenes}
\alias{landsat_load_scenes}
\title{Load multiple scenes from a vector of image filenames}
\usage{
landsat_load_scenes(path, include_bands = NULL)
}
\arguments{
\item{path}{A vector of landsat image filenames}

\item{include_bands}{A vector of band codes to include in the output,
or NULL to use all of them.}
}
\value{
A data.frame with the same columns as \link{landsat_parse_filename_image}
  plus one named \code{scene}, containing the landsat scenes as loaded by
  \link{landsat_scene}.
}
\description{
This function applies \link{landsat_scene} to a vector of image filenames, such
as that listed by \link{landsat_list_images}.
}
