% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_transform.R
\name{landsat_overlay}
\alias{landsat_overlay}
\title{Overlay a function on a Landsat object}
\usage{
landsat_overlay(x, fun, ...)
}
\arguments{
\item{x}{A \link{landsat_scene} or landsat_scene_df (from \link{landsat_load_scenes})}

\item{fun}{Function with formal arguments B1, B2, B3, etc.}

\item{...}{Passed to raster::\link[raster]{overlay}}
}
\value{
A landsat_scene or landsat_scene_df
}
\description{
This function makes it easy to calculate indicies such as NDWI, NDVI,
etc. on \link{landsat_scene} objects.
}
