% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_files.R
\name{landsat_parse_filename}
\alias{landsat_parse_filename}
\title{Extract Information from Landsat Filenames}
\usage{
landsat_parse_filename(path)
}
\arguments{
\item{path}{Filenames or file paths that are valid Landsat GeoTIFF files.}
}
\value{
A data.frame with the same number of rows as the length of \code{path}.
}
\description{
Uses Landsat filenames (that perhaps were listed using \link{landsat_list_files})
to fill in data that are present in the filename.
}
\references{
\url{https://landsat.usgs.gov/what-are-naming-conventions-landsat-scene-identifiers}
}
