% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_files.R
\name{landsat_parse_filename_image}
\alias{landsat_parse_filename_image}
\alias{landsat_parse_filename_archive}
\alias{landsat_parse_product_id}
\title{Extract Information from Landsat Filenames}
\usage{
landsat_parse_filename_image(path)

landsat_parse_filename_archive(path)

landsat_parse_product_id(path)
}
\arguments{
\item{path}{Filenames or file paths that are valid Landsat GeoTIFF files.}
}
\value{
A data.frame with the same number of rows as the length of \code{path}.
}
\description{
Uses Landsat filenames (that perhaps were listed using \link{landsat_list_images}
or \link{landsat_list_archives})
to fill in data that are present in the filename.
}
\examples{
landsat_parse_product_id("LT05_L1TP_063013_19960711_20170104_01_T1")
landsat_parse_filename_archive("LT05_L1TP_063013_19960711_20170104_01_T1.tar.gz")
landsat_parse_filename_image("LT05_L1TP_063013_19960711_20170104_01_T1_B1.TIF")

landsat_images <- landsat_list_images()
landsat_parse_filename_image(landsat_images)

landsat_archives <- landsat_list_archives()
landsat_parse_filename_archive(landsat_archives)

}
\references{
\url{https://landsat.usgs.gov/what-are-naming-conventions-landsat-scene-identifiers}
}
