% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_transform.R
\name{landsat_project}
\alias{landsat_project}
\title{Project Landsat objects}
\usage{
landsat_project(x, crs_obj)
}
\arguments{
\item{x}{A \link{landsat_scene} or landsat_scene_df (from \link{landsat_load_scenes})}

\item{crs_obj}{Usually an object of type
\link[sp]{CRS}, but could also be from the
sf package (\link[sf]{st_bbox}), an integer describing an EPSG code,
or a character string describing proj4 arguments.}
}
\value{
A landsat_scene or landsat_scene_df
}
\description{
This function projects a \link{landsat_scene} or landsat_scene_df
(as generated by \link{landsat_load_scenes}) using an object
describing a CRS This is usually an object of type
\link[sp]{CRS}, but could also be from the
sf package (\link[sf]{st_bbox}), an integer describing an EPSG code,
or a character string describing proj4 arguments.
}
