% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_scene.R
\name{landsat_scene}
\alias{landsat_scene}
\alias{as.landsat_scene}
\alias{is.landsat_scene}
\title{Create a RasterStack with attached Band Name Information}
\usage{
landsat_scene(B1 = NULL, B2 = NULL, B3 = NULL, B4 = NULL, B5 = NULL,
  B6 = NULL, B7 = NULL, B8 = NULL, B9 = NULL, B10 = NULL,
  B11 = NULL, BQA = NULL, attrs = list())

as.landsat_scene(x, band_names, attrs = list())

is.landsat_scene(x)
}
\arguments{
\item{B1, B2, B3, B4, B5, B6, B7, B8, B9, B10, B11, BQA}{\code{RasterLayer} or
filenames of raster image files to be passed to \link[raster]{raster}.}

\item{attrs}{A \code{list} that contains arbitrary meta information to keep
with the \code{landsat_scene} object.}

\item{x}{A RasterBrick or RasterStack object}

\item{band_names}{The band names associated with the RasterBrick or
Raster Stack object.}
}
\value{
A \code{RasterStack} object with band names attached as attributes.
}
\description{
This function creates a RasterStack (see \link[raster]{stack}) with band name
information attached. This information will be ignored by most functions in the
\code{raster} package, but is propogated by \link{landsat_crop}, \link{landsat_project},
\link{landsat_mask}, and \link{landsat_overlay}. The primary reason to use keep
the band name information with the RasterStack/RasterBrick object is for use with
\link{landsat_overlay}, which allows functions to be written in terms of band
numbers. This function does not load any raster files into memory.
}
\seealso{
\link{landsat_load_scenes}
}
