% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_transform.R
\name{landsat_transform}
\alias{landsat_transform}
\alias{landsat_transform.default}
\alias{landsat_transform.landsat_scene_list}
\alias{landsat_transform.landsat_scene_df}
\alias{landsat_transform.data.frame}
\title{Propogate Landsat attributes through transformations}
\usage{
landsat_transform(x, ...)

\method{landsat_transform}{default}(x, fun, ...)

\method{landsat_transform}{landsat_scene_list}(x, fun, ...)

\method{landsat_transform}{landsat_scene_df}(x, fun, ...)

\method{landsat_transform}{data.frame}(x, fun, ...)
}
\arguments{
\item{x}{A \link{landsat_scene} or landsat_scene_df (from \link{landsat_load_scenes})}

\item{...}{Arguments to be passed to \code{fun}}

\item{fun}{A funcion takes a RasterStack/RasterBrick and returns a
RasterStack/RasterBrick}
}
\value{
A  \link{landsat_scene}, RasterStack, RasterBrick, or landsat_scene_df
}
\description{
This function applies a raster transformation function (a function that takes
a RasterStack/RasterBrick and returns a RasterStack/RasterBrick) to a
\link{landsat_scene} or landsat_scene_df (from \link{landsat_load_scenes}).
}
